refactor: LottoInput -> LottoTicket 으로 변경
